+++ src/components/home/Card.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import styled from 'styled-components';
-import { toJapanDate } from '@/utils/helper/dayjs';
+import { toJapanDate } from '@/libs/dayjs';
 import { Post } from '@/utils/types/blog';
 
 export default function Card({ post }: { post: Post }) {
